Create a mediator registration flag and environment variable
Prior to this change, registration of the mediator with OpenHIM was
automatic, at mediator startup

This change introduces a configuration flag to indicate whether
to register this mediator with a running OpenHIM instance (true)
or not (false). The default value for the flag is true.

OHM-877
