+++ src/config.js
@@ -8,7 +8,8 @@ exports.getConfig = function() {
       apiURL: process.env.OPENHIM_URL || 'https://localhost:8080',
       username: process.env.OPENHIM_USERNAME || 'root@openhim.org',
       password: process.env.OPENHIM_PASSWORD || 'openhim-password',
-      trustSelfSigned: process.env.TRUST_SELF_SIGNED === 'true'
+      trustSelfSigned: process.env.TRUST_SELF_SIGNED === 'true',
+      register: process.env.OPENHIM_REGISTER || 'true'
     })
   })
 }
